Optimized json conversion

+++ csv-to-json-local-pyspark.py
@@ -1,8 +1,6 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Aug 12 13:59:32 2022
-
 @author: ditoj
 @references:
 https://www.youtube.com/watch?v=zhbMKX9eIPM
@@ -13,8 +11,12 @@ import shutil
 from pyspark.sql import SparkSession
 from pyspark.sql import functions as f
 
+
 def write_json(df, filepath):
-    shutil.rmtree(filepath)
+
+    if os.path.exists(filepath) and os.path.isdir(filepath):
+        shutil.rmtree(filepath)
+
     df.write.json(filepath)
     
     for root, dirs, files in os.walk(filepath):
@@ -28,42 +30,28 @@ spark = SparkSession.builder.appName('csv-to-json').getOrCreate()
 
 # Read CSV and create dataframe
 print("Reading CSV ...")
-csvdf = spark.read.csv('employee.csv', header=True, inferSchema=True)
+csvdf = spark.read.csv("employee.csv", header=True, inferSchema=True)
 csvdf.printSchema()
 csvdf.show()
 
 # Derive CSV fields
-print("Deriving fields ...")
-csvdf = csvdf.withColumn('RevSalary', csvdf.Salary*1.1)
-
-csvdf.printSchema()
-csvdf.show()
-
-# Convert to JSON dataframe
-print("Converting to JSON ...")
-print(csvdf.toJSON().collect())
-write_json(csvdf, "JSON")
-
-
-# Map to Nested JSON schema and Convert
 # https://stackoverflow.com/questions/58033947/create-nested-json-out-of-pyspark-dataframe
-print("Mapping and converting to Nested JSON ...")
-nestdf = (csvdf
-          .withColumn('Financials', f.struct(csvdf.Salary, csvdf.RevSalary))
-          .select('Name', 'Age', 'Financials'))
-
-print(nestdf.toJSON().collect())
-write_json(nestdf, "nestedJSON")
-
-# Lookup values
 # https://www.educba.com/pyspark-broadcast-join/
 # https://henning.kropponline.de/2016/12/11/broadcast-join-with-spark/
 # https://stackoverflow.com/questions/34053302/pyspark-and-broadcast-join-example
 # https://sparkbyexamples.com/pyspark/pyspark-broadcast-variables/
-print("Looking up values to Nested JSON ...")
-lkpdf = spark.read.csv('department.csv', header=True, inferSchema=True)
-lkpdf.printSchema()
-lkpdf.show()
 
-finaldf = csvdf.join(f.broadcast(lkpdf), csvdf.Dept == lkpdf.Dept).select('Name', 'Age', csvdf.Dept, 'Manager')
-write_json(finaldf, "lookupJSON")
+print("Deriving fields ...")
+lkpdf = spark.read.csv("department.csv", header=True, inferSchema=True)
+jsondf = (csvdf
+          .withColumn('RevSalary', csvdf.Salary * 1.1)  # Derive fields
+          .withColumn('Financials', f.struct('Salary', 'RevSalary'))  # Derive nested fields
+          .join(f.broadcast(lkpdf), csvdf.Dept == lkpdf.Dept)  # Lookup values
+          .select('Name', 'Age', csvdf.Dept, 'Manager', 'Financials')  # Select subset of fields
+          )
+jsondf.printSchema()
+jsondf.show()
+
+# Write to JSON
+print("Writing to JSON ...")
+write_json(jsondf, "csv-to-json")
